+++ spi/common/verifiable-credentials-spi/src/main/java/org/eclipse/edc/iam/verifiablecredentials/spi/model/VerifiableCredential.java
@@ -71,7 +71,7 @@ public class VerifiableCredential {
         return issuer;
     }
 
-    @JsonAlias({ "issued" })
+    @JsonAlias({ "issued" }) // some credentials like StatusList2021 don't adhere to the spec
     @NotNull
     public Instant getIssuanceDate() {
         return issuanceDate;
